chore: move calculateNewFileUri to the adapters
