+++ client/src/connection/rest/RestServerAdapter.ts
@@ -39,7 +39,7 @@ import {
 } from "./util";
 
 export const SAS_SERVER_HOME_DIRECTORY = "SAS_SERVER_HOME_DIRECTORY";
-export const SAS_FILE_SEPARATOR = "~fs~";
+const SAS_FILE_SEPARATOR = "~fs~";
 
 class RestServerAdapter implements ContentAdapter {
   protected baseUrl: string;
@@ -458,6 +458,97 @@ class RestServerAdapter implements ContentAdapter {
     } catch (error) {
       return;
     }
+  }
+
+  public calculateNewFileUri(
+    closedFileUri: Uri,
+    movedItem: ContentItem,
+    newItemUri: Uri,
+  ): Uri | null {
+    const isFolder = movedItem.fileStat?.type === FileType.Directory;
+
+    // If the moved item is a file and matches the closed file, return the new URI
+    if (
+      !isFolder &&
+      closedFileUri.toString() === movedItem.vscUri?.toString()
+    ) {
+      return newItemUri;
+    }
+
+    // If the moved item is a folder, calculate the new path for files within it
+    if (isFolder && movedItem.vscUri) {
+      const extractPathFromUri = (uri: string): string => {
+        try {
+          const queryStart = uri.indexOf("?");
+          if (queryStart === -1) {
+            return "";
+          }
+
+          const queryString = uri.substring(queryStart + 1);
+          const decodedQuery = decodeURIComponent(queryString);
+          const idMatch = decodedQuery.match(/id=(.+)/);
+          if (!idMatch || !idMatch[1]) {
+            return "";
+          }
+
+          // Extract the file path from the REST server URI format
+          // Removes the compute session prefix to get the actual file path
+          const uriWithoutPrefix = idMatch[1].replace(
+            /\/compute\/sessions\/[a-zA-Z0-9-]*\/files\//,
+            "",
+          );
+          try {
+            return decodeURIComponent(uriWithoutPrefix);
+          } catch (error) {
+            console.error("Failed to decode URI component:", error);
+            return uriWithoutPrefix;
+          }
+        } catch (error) {
+          console.error("Failed to extract path from URI:", error);
+          return "";
+        }
+      };
+
+      const oldBasePath = extractPathFromUri(movedItem.vscUri.toString());
+      const closedFilePath = extractPathFromUri(closedFileUri.toString());
+
+      // Check if the closed file was inside the moved folder
+      // Match if: closedFilePath starts with oldBasePath followed by separator
+      const isChildFile =
+        oldBasePath &&
+        closedFilePath &&
+        (closedFilePath.startsWith(oldBasePath + SAS_FILE_SEPARATOR) ||
+          closedFilePath === oldBasePath);
+
+      if (isChildFile && oldBasePath !== closedFilePath) {
+        try {
+          const relativePath = closedFilePath.substring(oldBasePath.length);
+          const filename = relativePath.replace(/^~fs~/, "");
+          const newUriStr = newItemUri.toString();
+
+          // Extract and modify the query to append the filename path
+          const queryMatch = newUriStr.match(/\?(.+)$/);
+          if (!queryMatch) {
+            return null;
+          }
+
+          const decodedQuery = decodeURIComponent(queryMatch[1]);
+          const newQuery = decodedQuery.replace(
+            /(\/files\/[^&]*)/,
+            `$1~fs~${filename}`,
+          );
+
+          return Uri.parse(
+            `${newItemUri.scheme}:/${filename}?${encodeURIComponent(newQuery)}`,
+          );
+        } catch (error) {
+          console.error("Failed to construct new file URI:", error);
+          return null;
+        }
+      }
+    }
+
+    return null;
   }
 
   public async renameItem(
